Rip out Gram matrix iteration and replace
The Gram matrix iteration process has been replaced with direct
variation of the component vectors; this results in a major speedup.

In addition, iterateOnDesign has an extra parameter for the variation of
the step size: the step size is now

  (errorMultiplier * error^errorExp) / (totalBadness + 1)

and after some casual experimentation it seems that errorExp ~ 1.1 and
erorMultiplier ~ 1 are good values.