Save badness in search_designs.m
Based on interesting observations (very small errors for a design that
should not exist, but corresponding to a very high badness compared
to searches with 'close' n), we now save the badness in search_designs.m
and we display it in the HTML from generate_from_search.py in the same
way as it is shown in generate_runtf.py (i.e. fractionally).